requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
autopep8==2.3.1
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.1
decorator==5.1.1
distro==1.9.0
Django==4.2
django-extensions==3.2.3
django-redis==5.4.0
django-seed==0.3.1
django-silk==5.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
Faker==28.0.0
Flask==3.0.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
gprof2dot==2024.6.6
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.6
inflection==0.5.1
ipython==8.26.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
openai==1.43.0
openpyxl==3.1.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.8
referencing==0.35.1
render==1.0.0
requests==2.31.0
rpds-py==0.20.0
Send2Trash==1.8.2
six==1.16.0
slack-bolt==1.18.1
slack_sdk==3.26.2
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.27
sqlparse==0.5.1
stack-data==0.6.3
tomli==2.0.1
toposort==1.10
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
